Migration cleanup
