+++ django_spire/notification/sms/processor.py
@@ -44,7 +44,7 @@ class SMSNotificationProcessor(BaseNotificationProcessor):
 
                 notification.status = NotificationStatusChoices.SENT
                 notification.sent_datetime = now()
-            except:
+            except Exception:
                 notification.status = NotificationStatusChoices.ERRORED
 
         Notification.objects.bulk_update(notifications, ['status', 'sent_datetime'])
